data_access: separate Body into its own file
There's a bit too much to pretend this is all details of APIProvider. Create a new interface, using it properly from OkHttpAPIProvider.
